Removed duplicated code of GDB-Parser
Now plugins for all version of Visual Studio share the same code.
