+++ src/GDBParser/GDBParserAPI.cpp
@@ -426,7 +426,7 @@ bool GDBParser::LaunchProcess(String^ pidStr, String^ exeStr, String^ IPAddrStr,
 	{
 		m_remotePath = m_remotePath + "\\lib;" + m_remotePath + "\\usr\\lib;" + m_remotePath + "\\usr\\lib\\qt4";
 		CAutoPtr <char> apPath = convertToAutoPtrFromString(m_remotePath);
-		sprintf(pcCmd, "6set solib-search-path %s\r\n", apPath );		
+		sprintf(pcCmd, "6set solib-search-path %s\r\n", apPath);    
 		console->sendCommand(pcCmd);
 		response = console->waitForPrompt(true);
 		parsed = parseGDB(response, parsingInstructions[8]);
@@ -436,7 +436,7 @@ bool GDBParser::LaunchProcess(String^ pidStr, String^ exeStr, String^ IPAddrStr,
 			return false;
 		}
 	}
-	
+  
 	sprintf(pcCmd, "7-target-attach %d\r\n", pid);
 	console->sendCommand(pcCmd);
 	response = console->waitForPrompt(true);
