+++ src/GDBParser/Instructions.txt
@@ -11,9 +11,9 @@ set:->:??^done;{#done;};
 -exec-arguments:->:??^done;#Arguments loaded.;
 -exec-run:->:??^running;{0;#40\;;??thread-created,id=";@@";0;??pid=";#\;;@@";0;#$EOL$;(??~";#80\;";?<?~;@@$EOL$;#"!80;):$EOL$;(??&";#80\;";?<?&;@@$EOL$;#"!80;):$EOL$;};
 -exec-continue:->:??^running;{#41\;;??thread-id="all;{#0;}{??thread-id=";@@";};0;#$EOL$;(??~";#80\;";?<?~;@@$EOL$;#"!80;):$EOL$;(??&";#80\;";?<?&;@@$EOL$;#"!80;):$EOL$;};
--exec-step:->:??^running;{#41\;;??thread-id="all;{#0;}{??thread-id=";@@";};0;#$EOL$;(??~";#80\;";?<?~;@@$EOL$;;#"!80;):$EOL$;(??&";#80\;";?<?&;@@$EOL$;;#"!80;):$EOL$;};
--exec-next:->:??^running;{#41\;;??thread-id="all;{#0;}{??thread-id=";@@";};0;#$EOL$;(??~";#80\;";?<?~;@@$EOL$;;#"!80;):$EOL$;(??&";#80\;";?<?&;@@$EOL$;;#"!80;):$EOL$;}{??^error;{#49\;;??msg=";@@";};};
--exec-finish:->:??^running;{#41\;;??thread-id="all;{#0;}{??thread-id=";@@";};0;#$EOL$;(??~";#80\;";?<?~;@@$EOL$;;#"!80;):$EOL$;(??&";#80\;";?<?&;@@$EOL$;;#"!80;):$EOL$;};
+-exec-step:->:??^running;{#41\;;??thread-id="all;{#0;}{??thread-id=";@@";};0;#$EOL$;(??~";#80\;";?<?~;@@$EOL$;#"!80;):$EOL$;(??&";#80\;";?<?&;@@$EOL$;#"!80;):$EOL$;};
+-exec-next:->:??^running;{#41\;;??thread-id="all;{#0;}{??thread-id=";@@";};0;#$EOL$;(??~";#80\;";?<?~;@@$EOL$;#"!80;):$EOL$;(??&";#80\;";?<?&;@@$EOL$;#"!80;):$EOL$;}{??^error;{#49\;;??msg=";@@";};};
+-exec-finish:->:??^running;{#41\;;??thread-id="all;{#0;}{??thread-id=";@@";};0;#$EOL$;(??~";#80\;";?<?~;@@$EOL$;#"!80;):$EOL$;(??&";#80\;";?<?&;@@$EOL$;#"!80;):$EOL$;};
 tbreak:->:??=breakpoint-created;{#20\;;??number=";{@@";};#\;;??enabled=";{@@";};#\;;??addr=";{@@";};#\;;??func=";{@@";};#\;;??file=";{@@";};#\;;??line=";{@@";};#\;;??times=";{@@";};}{??^error;{#ERROR: ;??msg=";@@";#$EOL$;(??&";#80\;";?<?&;@@$EOL$;#"!80;):$EOL$;$END$;};#tbreak: Function not found!;};
 -break-insert:->:??^done;{#20\;;??number=";{@@";};#\;;??enabled=";{@@";};#\;;??addr=";{@@";};#\;;??func=";{@@";};#\;;??file=";{@@";};#\;;??line=";{@@";};#\;;??times=";{@@";};}{#Function not found!;};
 -break-watch:->:??^done;{#20\;;??number=";{@@";};#\;;??enabled=";{@@";};#\;;??addr=";{@@";};#\;;??func=";{@@";};#\;;??file=";{@@";};#\;;??line=";{@@";};#\;;??times=";{@@";};}{#Function not found!;};
