--- src_vs2012/GDBParser/Instructions.txt
@@ -1,48 +1,0 @@
-asynchronous:->:??=thread-exited,id;{#51\;;??";@@";#$EOL$;};??thread-created,id;{#40\;;??";@@";#\;;??pid ;@@ ;#$EOL$;};0;??*running;{#41\;;??thread-id="all;{#0;}{??thread-id=";@@";};#$EOL$;};0;??*stopped;{??breakpoint-hit;{#27\;;??bkptno=";@@";#\;;??file=";@@";#\;;??line=";@@";}{??end-stepping-range;{#45;??file=";{#\;;@@";0;#\;;??line=";@@";};}{??function-finished;{#46;??file=;{#\;;??";@@";#\;;??line=";@@";};}{??exited-normally;{#42;}{??exit-code;{#43\;;??";@@";}{??signal-received;{??signal-meaning="Killed";{#47;}{??signal-meaning="Segmentation fault";{#54\;;??addr=";@@";#\;;??func=";@@";0;??file=;{#\;;??";@@";#\;;??line=";@@";};}{??signal-meaning="Aborted";{#55\;;0;??signal-name=";@@";#\;;??signal-meaning=";@@";}{#44\;;??addr=";@@";#\;;??func=";@@";0;??file=;{#\;;??";@@";#\;;??line=";@@";};};};};}{??exited-signalled;{#55\;;0;??signal-name=";@@";#\;;??signal-meaning=";@@";}{#44\;;??addr=";@@";#\;;??func=";@@";0;??file=;{#\;;??";@@";#\;;??line=";@@";};};};};};};};};#\;;??thread-id=;{??";@@";};#$EOL$;};0;??=breakpoint-modified,bkpt=\{number=";{0;($EOR$;$1=?<?=breakpoint-modified,bkpt=\{number=";@@";$$;$EOR$;?<?=breakpoint-modified,bkpt=\{number="$1$;#21\;$1$\;;??enabled=";@@";#\;;??addr=";@@";#\;;??func=";@@";#\;;??file=";@@";#\;;??line=";@@";#\;;??times=";@@";0;(??=breakpoint-modified,bkpt=\{number="$1$;%?<?$EOL$;@@"\};);):$EOL$;};0;??=breakpoint-deleted,id=;{0;(??=breakpoint-deleted,id=";#22\;;@@";):$EOL$;};0;??@";{0;#81\;";(??@";?<?@;@@$EOL$;):$EOL$;#"!81$EOL$;};0;(??~";#80\;";?<?~;@@$EOL$;#"!80;):$EOL$;??Error in testing breakpoint condition;{#29\;$EOL$;};??Quit (expect signal SIGINT when the program is resumed);{#50\;$EOL$;};??2374: internal-error: frame_cleanup_after_sniffer: Assertion;{#52\;$EOL$;};??^error,msg="Remote communication error: No error.;{#53\;$EOL$;};??: internal-error: handle_inferior_event: Assertion;{#56\;$EOL$;};0;(??&";#80\;";?<?&;@@$EOL$;#"!80;):$EOL$;(??$EOL$=;#80\;";?<?=;@@$EOL$;#"!80;):$EOL$;
-=breakpoint-modified:->:($EOR$;$1=?<?=breakpoint-modified,bkpt=\{number=";@@";$$;$EOR$;?<?=breakpoint-modified,bkpt=\{number="$1$;#21\;$1$\;;??enabled=";@@";#\;;??addr=";@@";#\;;??func=";@@";#\;;??file=";@@";#\;;??line=";@@";#\;;??times=";@@";0;(??=breakpoint-modified,bkpt=\{number="$1$;%?<?$EOL$;@@"\};);):$EOL$;
-starting:->:??(gdb);{0;#80\;";(??~";?<?~;@@$EOL$;):$EOL$;#"!80;}{#!console=ERROR: Could not load GDB!};
--target-select:->:??^connected;{0;#80\;";(??~";?<?~;@@$EOL$;):$EOL$;#"!80;}{#!console=ERROR: Could not set target IP:Port};
-file:->:??^done;{0;#80\;";(??~";?<?~;@@$EOL$;):$EOL$;#"!80;};
-upload:->:??^done;#File uploaded!;
--target-attach:->:??^done;{#state="stopped"\;processID=";0;??pid=";@@";#"\;thread-ID=";??thread-created,id=";@@";0;#"\;$EOL$80\;";(??~";?<?~;@@$EOL$;):$EOL$;#"!80;};
--file-exec-and-symbols:->:??^done;#File and symbols loaded.;
--gdb-set:->:??^done;{#done;};
-set:->:??^done;{#done;};
--exec-arguments:->:??^done;#Arguments loaded.;
--exec-run:->:??^running;{0;#40\;;??thread-created,id=";@@";0;??pid=";#\;;@@";0;#$EOL$;(??~";#80\;";?<?~;@@$EOL$;#"!80;):$EOL$;(??&";#80\;";?<?&;@@$EOL$;#"!80;):$EOL$;};
--exec-continue:->:??^running;{#41\;;??thread-id="all;{#0;}{??thread-id=";@@";};0;#$EOL$;(??~";#80\;";?<?~;@@$EOL$;#"!80;):$EOL$;(??&";#80\;";?<?&;@@$EOL$;#"!80;):$EOL$;};
--exec-step:->:??^running;{#41\;;??thread-id="all;{#0;}{??thread-id=";@@";};0;#$EOL$;(??~";#80\;";?<?~;@@$EOL$;#"!80;):$EOL$;(??&";#80\;";?<?&;@@$EOL$;#"!80;):$EOL$;};
--exec-next:->:??^running;{#41\;;??thread-id="all;{#0;}{??thread-id=";@@";};0;#$EOL$;(??~";#80\;";?<?~;@@$EOL$;#"!80;):$EOL$;(??&";#80\;";?<?&;@@$EOL$;#"!80;):$EOL$;}{??^error;{#49\;;??msg=";@@";};};
--exec-finish:->:??^running;{#41\;;??thread-id="all;{#0;}{??thread-id=";@@";};0;#$EOL$;(??~";#80\;";?<?~;@@$EOL$;#"!80;):$EOL$;(??&";#80\;";?<?&;@@$EOL$;#"!80;):$EOL$;};
-tbreak:->:??=breakpoint-created;{#20\;;??number=";{@@";};#\;;??enabled=";{@@";};#\;;??addr=";{@@";};#\;;??func=";{@@";};#\;;??file=";{@@";};#\;;??line=";{@@";};#\;;??times=";{@@";};}{??^error;{#ERROR: ;??msg=";@@";#$EOL$;(??&";#80\;";?<?&;@@$EOL$;#"!80;):$EOL$;$END$;};#tbreak: Function not found!;};
--break-insert:->:??^done;{#20\;;??number=";{@@";};#\;;??enabled=";{@@";};#\;;??addr=";{@@";};#\;;??func=";{@@";};#\;;??file=";{@@";};#\;;??line=";{@@";};#\;;??times=";{@@";};}{#Function not found!;};
--break-watch:->:??^done;{#20\;;??number=";{@@";};#\;;??enabled=";{@@";};#\;;??addr=";{@@";};#\;;??func=";{@@";};#\;;??file=";{@@";};#\;;??line=";{@@";};#\;;??times=";{@@";};}{#Function not found!;};
-$-break-delete:->:??^done;{#25;@@$EOL$};
-$-break-disable:->:??^done;{#24;@@$EOL$};
-$-break-enable:->:??^done;{#23;@@$EOL$};
-$-break-after:->:??^done;{#26;@@$EOL$};
-$-break-condition:->:??^done;{#28;@@$EOL$};
--break-list:->:??^done;{(??number=";#21\;;@@";#\;;??enabled=";@@";#\;;??addr=";@@";#\;;??func=";@@";#\;;??file=";@@";#\;;??line=";@@";#\;;??times=";@@";):$EOL$;};
-kill:->:??^done;{#done;};
-generate-core-file:->:#NOT CREATED!
-jump:->:#NOT CREATED!
-handle:->:#NOT CREATED!
-watch:->:#NOT CREATED!
-interrupt:->:#NOT CREATED!
-show:->:#NOT CREATED!
--stack-info-depth:->:??^done,depth=";@@";$EOL$
--stack-list-frames:->:??^done;{(??level=;{??";@@";};~"\};#\;;??addr=;{??";@@";};#\;;??func=;{??";@@";};#\;;??file=;{??";@@";};#\;;??line=;{??";@@";};);};
--stack-select-frame:->:??^done;{#done;};
--stack-list-variables:->:??^done;{#70\;;(??name=;{??";@@";};~"\};#\;;??arg=;{??";@@";};#\;;??type=;{??";@@";};#\;;??value=;{??";@@";};);}{#ERROR;};
--var-create:->:??^done;{??numchild=";@@";}{#ERROR;};
--var-delete:->:??^done;{#DONE;}{#ERROR;};
--var-list-children:->:??^done;{#71\;;(??name=;{??";@@";};~"\};#\;;??numchild=;{??";@@";};#\;;??value=;{??";@@";};#\;;??type=;{??";@@";};);}{#ERROR;};
--data-evaluate-expression:->:??^done;{??=;#60\;;@@$EOL$;}{??^error;{??-data-evaluate-expression: Usage: -data-evaluate-expression expression;{#61\;Error in expression.;}{??A syntax error in expression, near `'.;{#61\;A syntax error at the end of the expression.;}{??=;#61\;;@@$EOL$;};};};};
-ptype:->:??type = ;{@@\\n";}{# ;};
-whatis:->:??type = ;{@@\\n";}{# ;};
--exec-interrupt:->:??^done;{??signal-received;{??signal-meaning="Killed";{#47;}{#44\;;??addr=";@@";#\;;??func=";@@";0;??file=;{#\;;??";@@";#\;;??line=";@@";};};};};#\;;??thread-id=;{??";@@";};
--target-detach:->:#NOT CREATED!
--var-info-num-children:->:??numchild;{??";@@";}{#ERROR;};
-info:->:??Address;{??0x;?<? ;@@ ;}{??Line;{@@of;};};
--thread-info:->:??^done;{(??\{id=";@@";#\;;??state=;{??";@@";};#\;;~\};?<?target-id=;{??";@@",;};#\;;??details=;{??";@@";};#\;;??fullname=;{??";@@";#\;;??line=";@@";#\;;?<?frame=;}{??frame=;};?<?name=;{??";@@";}{?<?name \\";{@@\\"";};};?<?\{id=";%?<?\{;@@state=";);0;??current-thread-id=";@@";};
--thread-select:->:??^done;{#done;};
