--- src_vs2012/GDBParser/ProjInclude.h
@@ -1,33 +1,0 @@
-//* Copyright 2010-2011 Research In Motion Limited.
-//*
-//* Licensed under the Apache License, Version 2.0 (the "License");
-//* you may not use this file except in compliance with the License.
-//* You may obtain a copy of the License at
-//*
-//* http://www.apache.org/licenses/LICENSE-2.0
-//*
-//* Unless required by applicable law or agreed to in writing, software
-//* distributed under the License is distributed on an "AS IS" BASIS,
-//* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-//* See the License for the specific language governing permissions and
-//* limitations under the License.
-
-#pragma once
-
-using namespace System;
-using namespace System::Diagnostics;
-#include <vcclr.h>
-#pragma warning(push)
-#pragma warning(disable:4091)
-#include <msclr\lock.h>
-#pragma warning(pop)
-
-#define ASSERT(cond) _ASSERT(cond)
-#ifdef DEBUG
-	#define VERIFY(statement) Debug::Assert((statement) != 0, #statement)
-#else
-	#define VERIFY(statement) statement
-#endif
-
-#include "ComponentException.h"
-#include "WorkerUtil.h"
